requirements.txt
numpy
opencv-python
imageio
websockets
rerun-sdk
torch
